+++ klay_beam/src/klay_beam/torch_utils.py
@@ -7,6 +7,7 @@ def torch_check() -> Tuple[bool, Optional[ImportError]]:
     try:
         import torch
         import torchaudio
+
         torch_available = True
     except ImportError as e:
         torch_import_error = e